backstage/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
dnspython==2.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
future==0.18.2
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
openapi-codec==1.3.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
prettytable==3.6.0
prompt-toolkit==3.0.38
pycparser==2.21
pyecharts==2.0.2
PyJWT==2.6.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-whois==0.8.0
pytz==2022.7
redis==3.5.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.8.2
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wsproto==1.2.0